gestion plano, falta corregir cliparea al max y min

+++ trunk/src/linea.h
@@ -35,7 +35,7 @@ public:
     double getX2();
     double getY2();
     string getCapa();
-    void escalar(int escala);
+    void escalar(double escala);
 private:
     double x1,y1,x2,y2; 
     string capa;
